Extract the visual reader view controller
