+++ TestApp/Sources/Reader/Common/VisualReaderViewController.swift
@@ -1,0 +1,339 @@
+//
+//  Copyright 2023 Readium Foundation. All rights reserved.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
+//
+
+import Combine
+import Foundation
+import R2Navigator
+import R2Shared
+import SwiftUI
+import UIKit
+
+/// Base class for the reader view controller of a `VisualNavigator`.
+class VisualReaderViewController<N: UIViewController & Navigator>: ReaderViewController<N>, VisualNavigatorDelegate {
+    private(set) var stackView: UIStackView!
+    private lazy var positionLabel = UILabel()
+
+    private let ttsViewModel: TTSViewModel?
+    private let ttsControlsViewController: UIHostingController<TTSControls>?
+
+    init(
+        navigator: N,
+        publication: Publication,
+        bookId: Book.Id,
+        books: BookRepository,
+        bookmarks: BookmarkRepository,
+        highlights: HighlightRepository?
+    ) {
+        self.highlights = highlights
+
+        ttsViewModel = TTSViewModel(navigator: navigator, publication: publication)
+        ttsControlsViewController = ttsViewModel.map { UIHostingController(rootView: TTSControls(viewModel: $0)) }
+
+        super.init(
+            navigator: navigator,
+            publication: publication,
+            bookId: bookId,
+            books: books,
+            bookmarks: bookmarks
+        )
+
+        addHighlightDecorationsObserverOnce()
+        updateHighlightDecorations()
+
+        NotificationCenter.default.addObserver(self, selector: #selector(voiceOverStatusDidChange), name: UIAccessibility.voiceOverStatusDidChangeNotification, object: nil)
+    }
+
+    @available(*, unavailable)
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    deinit {
+        NotificationCenter.default.removeObserver(self)
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .white
+
+        updateNavigationBar(animated: false)
+
+        stackView = UIStackView(frame: view.bounds)
+        stackView.distribution = .fill
+        stackView.axis = .vertical
+        view.addSubview(stackView)
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        let topConstraint = stackView.topAnchor.constraint(equalTo: view.topAnchor)
+        // `accessibilityTopMargin` takes precedence when VoiceOver is enabled.
+        topConstraint.priority = .defaultHigh
+        NSLayoutConstraint.activate([
+            topConstraint,
+            stackView.rightAnchor.constraint(equalTo: view.rightAnchor),
+            stackView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            stackView.leftAnchor.constraint(equalTo: view.leftAnchor),
+        ])
+
+        addChild(navigator)
+        stackView.addArrangedSubview(navigator.view)
+        navigator.didMove(toParent: self)
+
+        stackView.addArrangedSubview(accessibilityToolbar)
+
+        positionLabel.translatesAutoresizingMaskIntoConstraints = false
+        positionLabel.font = .systemFont(ofSize: 12)
+        positionLabel.textColor = .darkGray
+        view.addSubview(positionLabel)
+        NSLayoutConstraint.activate([
+            positionLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            positionLabel.bottomAnchor.constraint(equalTo: navigator.view.bottomAnchor, constant: -20),
+        ])
+
+        if let state = ttsViewModel?.$state, let controls = ttsControlsViewController {
+            controls.view.backgroundColor = .clear
+
+            addChild(controls)
+            controls.view.translatesAutoresizingMaskIntoConstraints = false
+            view.addSubview(controls.view)
+            NSLayoutConstraint.activate([
+                controls.view.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+                controls.view.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 20),
+            ])
+            controls.didMove(toParent: self)
+
+            state
+                .sink { state in
+                    controls.view.isHidden = !state.showControls
+                }
+                .store(in: &subscriptions)
+        }
+    }
+
+    // MARK: - Navigation bar
+
+    private var navigationBarHidden: Bool = true {
+        didSet {
+            updateNavigationBar()
+        }
+    }
+
+    override func makeNavigationBarButtons() -> [UIBarButtonItem] {
+        var buttons: [UIBarButtonItem] = super.makeNavigationBarButtons()
+
+        // Text to speech
+        if let ttsViewModel = ttsViewModel {
+            buttons.append(UIBarButtonItem(image: UIImage(systemName: "speaker.wave.2.fill"), style: .plain, target: ttsViewModel, action: #selector(TTSViewModel.start)))
+        }
+
+        return buttons
+    }
+
+    func toggleNavigationBar() {
+        navigationBarHidden = !navigationBarHidden
+    }
+
+    func updateNavigationBar(animated: Bool = true) {
+        let hidden = navigationBarHidden && !UIAccessibility.isVoiceOverRunning
+        navigationController?.setNavigationBarHidden(hidden, animated: animated)
+        setNeedsStatusBarAppearanceUpdate()
+    }
+
+    override var preferredStatusBarUpdateAnimation: UIStatusBarAnimation {
+        .slide
+    }
+
+    override var prefersStatusBarHidden: Bool {
+        navigationBarHidden && !UIAccessibility.isVoiceOverRunning
+    }
+
+    // MARK: - VisualNavigatorDelegate
+
+    override func navigator(_ navigator: Navigator, locationDidChange locator: Locator) {
+        super.navigator(navigator, locationDidChange: locator)
+
+        positionLabel.text = {
+            if let position = locator.locations.position {
+                return "\(position) / \(publication.positions.count)"
+            } else if let progression = locator.locations.totalProgression {
+                return "\(progression)%"
+            } else {
+                return nil
+            }
+        }()
+    }
+
+    func navigator(_ navigator: VisualNavigator, didTapAt point: CGPoint) {
+        // Turn pages when tapping the edge of the screen.
+        guard !DirectionalNavigationAdapter(navigator: navigator).didTap(at: point) else {
+            return
+        }
+        // clear a current search highlight
+        if let decorator = self.navigator as? DecorableNavigator {
+            decorator.apply(decorations: [], in: "search")
+        }
+
+        toggleNavigationBar()
+    }
+
+    func navigator(_ navigator: VisualNavigator, didPressKey event: KeyEvent) {
+        // Turn pages when pressing the arrow keys.
+        DirectionalNavigationAdapter(navigator: navigator).didPressKey(event: event)
+    }
+
+    // MARK: - Highlights
+
+    private let highlights: HighlightRepository?
+    private var highlightContextMenu: UIHostingController<HighlightContextMenu>?
+    private let highlightDecorationGroup = "highlights"
+    private var currentHighlightCancellable: AnyCancellable?
+
+    func saveHighlight(_ highlight: Highlight) {
+        guard let highlights = highlights else { return }
+
+        Task {
+            do {
+                try await highlights.add(highlight)
+                toast(NSLocalizedString("reader_highlight_success_message", comment: "Success message when adding a bookmark"), on: view, duration: 1)
+            } catch {
+                print(error)
+                toast(NSLocalizedString("reader_highlight_failure_message", comment: "Error message when adding a new bookmark failed"), on: view, duration: 2)
+            }
+        }
+    }
+
+    func updateHighlight(_ highlightID: Highlight.Id, withColor color: HighlightColor) {
+        guard let highlights = highlights else { return }
+
+        Task {
+            try! await highlights.update(highlightID, color: color)
+        }
+    }
+
+    func deleteHighlight(_ highlightID: Highlight.Id) {
+        guard let highlights = highlights else { return }
+
+        Task {
+            try! await highlights.remove(highlightID)
+        }
+    }
+
+    private func addHighlightDecorationsObserverOnce() {
+        if highlights == nil { return }
+
+        if let decorator = navigator as? DecorableNavigator {
+            decorator.observeDecorationInteractions(inGroup: highlightDecorationGroup) { [weak self] event in
+                self?.activateDecoration(event)
+            }
+        }
+    }
+
+    private func updateHighlightDecorations() {
+        guard let highlights = highlights else { return }
+
+        highlights.all(for: bookId)
+            .assertNoFailure()
+            .sink { [weak self] highlights in
+                if let self = self, let decorator = self.navigator as? DecorableNavigator {
+                    let decorations = highlights.map { Decoration(id: $0.id, locator: $0.locator, style: .highlight(tint: $0.color.uiColor, isActive: false)) }
+                    decorator.apply(decorations: decorations, in: self.highlightDecorationGroup)
+                }
+            }
+            .store(in: &subscriptions)
+    }
+
+    private func activateDecoration(_ event: OnDecorationActivatedEvent) {
+        guard let highlights = highlights else { return }
+
+        currentHighlightCancellable = highlights.highlight(for: event.decoration.id).sink { _ in
+        } receiveValue: { [weak self] highlight in
+            guard let self = self else { return }
+            self.activateDecoration(for: highlight, on: event)
+        }
+    }
+
+    private func activateDecoration(for highlight: Highlight, on event: OnDecorationActivatedEvent) {
+        if highlightContextMenu != nil {
+            highlightContextMenu?.removeFromParent()
+        }
+
+        let menuView = HighlightContextMenu(colors: [.red, .green, .blue, .yellow],
+                                            systemFontSize: 20)
+
+        menuView.selectedColorPublisher.sink { [weak self] color in
+            self?.currentHighlightCancellable?.cancel()
+            self?.updateHighlight(event.decoration.id, withColor: color)
+            self?.highlightContextMenu?.dismiss(animated: true, completion: nil)
+        }
+        .store(in: &subscriptions)
+
+        menuView.selectedDeletePublisher.sink { [weak self] _ in
+            self?.currentHighlightCancellable?.cancel()
+            self?.deleteHighlight(event.decoration.id)
+            self?.highlightContextMenu?.dismiss(animated: true, completion: nil)
+        }
+        .store(in: &subscriptions)
+
+        highlightContextMenu = UIHostingController(rootView: menuView)
+
+        highlightContextMenu!.preferredContentSize = menuView.preferredSize
+        highlightContextMenu!.modalPresentationStyle = .popover
+
+        if let popoverController = highlightContextMenu!.popoverPresentationController {
+            popoverController.permittedArrowDirections = .down
+            popoverController.sourceRect = event.rect ?? .zero
+            popoverController.sourceView = view
+            popoverController.backgroundColor = .cyan
+            popoverController.delegate = self
+            present(highlightContextMenu!, animated: true, completion: nil)
+        }
+    }
+
+    // MARK: - Accessibility
+
+    /// Constraint used to shift the content under the navigation bar, since it is always visible when VoiceOver is running.
+    private lazy var accessibilityTopMargin: NSLayoutConstraint = self.stackView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor)
+
+    private lazy var accessibilityToolbar: UIToolbar = {
+        func makeItem(_ item: UIBarButtonItem.SystemItem, label: String? = nil, action: UIKit.Selector? = nil) -> UIBarButtonItem {
+            let button = UIBarButtonItem(barButtonSystemItem: item, target: (action != nil) ? self : nil, action: action)
+            button.accessibilityLabel = label
+            return button
+        }
+
+        let toolbar = UIToolbar(frame: .zero)
+        toolbar.items = [
+            makeItem(.flexibleSpace),
+            makeItem(.rewind, label: NSLocalizedString("reader_backward_a11y_label", comment: "Accessibility label to go backward in the publication"), action: #selector(goBackward)),
+            makeItem(.flexibleSpace),
+            makeItem(.fastForward, label: NSLocalizedString("reader_forward_a11y_label", comment: "Accessibility label to go forward in the publication"), action: #selector(goForward)),
+            makeItem(.flexibleSpace),
+        ]
+        toolbar.isHidden = !UIAccessibility.isVoiceOverRunning
+        return toolbar
+    }()
+
+    private var isVoiceOverRunning = UIAccessibility.isVoiceOverRunning
+
+    @objc private func voiceOverStatusDidChange() {
+        let isRunning = UIAccessibility.isVoiceOverRunning
+        // Avoids excessive settings refresh when the status didn't change.
+        guard isVoiceOverRunning != isRunning else {
+            return
+        }
+        isVoiceOverRunning = isRunning
+        accessibilityTopMargin.isActive = isRunning
+        accessibilityToolbar.isHidden = !isRunning
+        updateNavigationBar()
+    }
+
+    @objc private func goBackward() {
+        navigator.goBackward()
+    }
+
+    @objc private func goForward() {
+        navigator.goForward()
+    }
+}
