+++ TestApp/Sources/Reader/EPUB/EPUBViewController.swift
@@ -7,15 +7,17 @@
 import R2Navigator
 import R2Shared
 import ReadiumAdapterGCDWebServer
+import SwiftSoup
 import SwiftUI
 import UIKit
+import WebKit
 
 public extension FontFamily {
     // Example of adding a custom font embedded in the application.
     static let literata: FontFamily = "Literata"
 }
 
-class EPUBViewController: ReaderViewController<EPUBNavigatorViewController> {
+class EPUBViewController: VisualReaderViewController<EPUBNavigatorViewController> {
     private let preferencesStore: AnyUserPreferencesStore<EPUBPreferences>
 
     init(
@@ -85,14 +87,6 @@ class EPUBViewController: ReaderViewController<EPUBNavigatorViewController> {
         }
     }
 
-    override var currentBookmark: Bookmark? {
-        guard let locator = navigator.currentLocation else {
-            return nil
-        }
-
-        return Bookmark(bookId: bookId, locator: locator)
-    }
-
     @objc func highlightSelection() {
         if let selection = navigator.currentSelection {
             let highlight = Highlight(bookId: bookId, locator: selection.locator, color: .yellow)
@@ -100,9 +94,68 @@ class EPUBViewController: ReaderViewController<EPUBNavigatorViewController> {
             navigator.clearSelection()
         }
     }
+
+    // MARK: - Footnotes
+
+    private func presentFootnote(content: String, referrer: String?) -> Bool {
+        var title = referrer
+        if let t = title {
+            title = try? clean(t, .none())
+        }
+        if !suitableTitle(title) {
+            title = nil
+        }
+
+        let content = (try? clean(content, .none())) ?? ""
+        let page =
+            """
+            <html>
+                <head>
+                    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+                </head>
+                <body>
+                    \(content)
+                </body>
+            </html>
+            """
+
+        let wk = WKWebView()
+        wk.loadHTMLString(page, baseURL: nil)
+
+        let vc = UIViewController()
+        vc.view = wk
+        vc.navigationItem.title = title
+        vc.navigationItem.leftBarButtonItem = BarButtonItem(barButtonSystemItem: .done, actionHandler: { _ in
+            vc.dismiss(animated: true, completion: nil)
+        })
+
+        let nav = UINavigationController(rootViewController: vc)
+        nav.modalPresentationStyle = .formSheet
+        present(nav, animated: true, completion: nil)
+
+        return false
+    }
+
+    /// This regex matches any string with at least 2 consecutive letters (not limited to ASCII).
+    /// It's used when evaluating whether to display the body of a noteref referrer as the note's title.
+    /// I.e. a `*` or `1` would not be used as a title, but `on` or `好書` would.
+    private lazy var noterefTitleRegex: NSRegularExpression =
+        try! NSRegularExpression(pattern: "[\\p{Ll}\\p{Lu}\\p{Lt}\\p{Lo}]{2}")
+
+    /// Checks to ensure the title is non-nil and contains at least 2 letters.
+    private func suitableTitle(_ title: String?) -> Bool {
+        guard let title = title else { return false }
+        let range = NSRange(location: 0, length: title.utf16.count)
+        let match = noterefTitleRegex.firstMatch(in: title, range: range)
+        return match != nil
+    }
 }
 
-extension EPUBViewController: EPUBNavigatorDelegate {}
+extension EPUBViewController: EPUBNavigatorDelegate {
+    func navigator(_ navigator: Navigator, shouldNavigateToNoteAt link: R2Shared.Link, content: String, referrer: String?) -> Bool {
+        presentFootnote(content: content, referrer: referrer)
+    }
+}
 
 extension EPUBViewController: UIGestureRecognizerDelegate {
     func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldRecognizeSimultaneouslyWith otherGestureRecognizer: UIGestureRecognizer) -> Bool {
